Changed order of locking
We want the return forks function to return forks first

then we can lock out for status updates, this will save on deadlocks because of someone who locks out the return forks but waits for the chopsticks, since the chopstick return is behind the lock. It doesn't necessarily need to be though, since we are concerned about status edits, not chopstick edits.